Reworked world gen

+++ src/generation.h
@@ -1,5 +1,5 @@
 #pragma once
 
 
-double get_noise_tree(int x, int y);
-double get_noise(int x, int y);
+float get_noise_tree(int x, int y);
+float get_noise(int x, int y);
